Fix the notifier list processing
If the notifiers was defined explicitly with '-n' or '--notifiers'
option, then 'enabled' list could be non-empty first, but when all
notifiers are removed in the processing loop, it became empty, although
some available notifiers may be still unprocessed. In this case
'needInit' will get wrong value if the 'enabled.isEmpty()' will be used
for its calculation.